+++ csatar-plugins/csatar/lang/en/lang.php
@@ -564,6 +564,14 @@ return [
                 'active' => 'Active',
             ],
         ],
+        'locations' => [
+            'country' => 'Country',
+            'county' => 'County',
+            'city' => 'City',
+            'street_type' => 'Street Type',
+            'street' => 'Street',
+            'code' => 'Zip code',
+        ],
         'component' => [
             'general' => [
                 'validationExceptions' => [
@@ -637,6 +645,7 @@ return [
                     'teamReportAlreadyExists' => 'The team report for this team and year already exists.',
                     'teamReportCannotBeFound' => 'The team report cannot be found.',
                     'teamCannotBeFound' => 'The team cannot be found.',
+                    'missingLegalRelationship' => 'Legal relationship is not set for :name'
                 ],
             ],
             'teamReports' => [