+++ csatar-plugins/csatar/lang/hu/lang.php
@@ -561,6 +561,14 @@
                 'active' => 'Érvényes',
             ],
         ],
+        'locations' => [
+            'country' => 'Ország',
+            'county' => 'Megye',
+            'city' => 'Város',
+            'street_type' => 'Utca típus',
+            'street' => 'Utca',
+            'code' => 'Irányitószám',
+        ],
         'component' => [
             'general' => [
                 'validationExceptions' => [
@@ -634,6 +642,7 @@
                     'teamReportAlreadyExists' => 'Már létezik csapatjelentés e csapat számára, erre az évre.',
                     'teamReportCannotBeFound' => 'A csapatjelentés nem található.',
                     'teamCannotBeFound' => 'A csapat nem található.',
+                    'missingLegalRelationship' => ':name tagnak nincs beállítva a jogviszonya.'
                 ],
             ],
             'teamReports' => [
@@ -679,7 +688,7 @@
                 'twoFactorAuthFailed' => 'A hitelesítés sikertelen, kérlek próbáld újra!',
                 'twoFactorAuthSuccess' => 'A hitelesítés sikeres, köszönjük!',
                 'reset' => 'Két faktoros hitelesítés visszaállítása',
-                'resetSuccess' => 'A két faktoros hitelesítés skeresen vissza lett állítva. A felhasználónak a hitelesítő alkalmazásból is törölnie kell a fióokot.',
+                'resetSuccess' => 'A két faktoros hitelesítés sikeresen vissza lett állítva. A felhasználónak a hitelesítő alkalmazásból is törölnie kell a fiókot.',
             ],
             'accidentLog' => [
                 'accidentLog' => 'Baleseti napló',