+++ csatar-plugins/csatar/models/Locations.php
@@ -1,0 +1,39 @@
+<?php namespace Csatar\Csatar\Models;
+
+use Model;
+
+/**
+ * Model
+ */
+class Locations extends Model
+{
+    use \October\Rain\Database\Traits\Validation;
+
+    /*
+     * Disable timestamps by default.
+     * Remove this line if timestamps are defined in the database table.
+     */
+    public $timestamps = false;
+
+
+    /**
+     * @var string The database table used by the model.
+     */
+    public $table = 'csatar_csatar_locations';
+
+    protected $fillable = [
+        'country',
+        'county',
+        'city',
+        'street_type',
+        'street',
+        'number',
+        'code',
+    ];
+
+    /**
+     * @var array Validation rules
+     */
+    public $rules = [
+    ];
+}