+++ csatar-plugins/csatar/models/Mandate.php
@@ -277,7 +277,14 @@ class Mandate extends Model
      */
     public function scopeMandateModelType($query, $model = null)
     {
-        return $model ? $query->where('mandate_model_type', $model::getModelName()) : $query->whereNull('id');
+        $currentDate = (new DateTime())->format('Y-m-d');
+        return $model
+            ? $query->where('mandate_model_type', $model::getModelName())
+                ->where('start_date', '<=', $currentDate)
+                ->where(function ($query) use ($currentDate) {
+                    return $query->whereNull('end_date')->orWhere('end_date', '>=', $currentDate);
+                })
+            : $query->whereNull('id');
     }
 
     public function getMandateTeamAttribute(): string
@@ -365,6 +372,17 @@ class Mandate extends Model
         return $query->whereHas('mandate_type', function ($query) use ($associationIds) {
             $query->whereIn('association_id', $associationIds);
         });
+    }
+
+    public function scopeInactiveMandatesInOrganizations($query, $record) {
+        $currentDate = (new DateTime())->format('Y-m-d');
+        $query->where('mandate_model_type', $record::getModelName())
+            ->where('mandate_model_id', $record->id)
+            ->where(function ($query) use ($currentDate) {
+                return $query->where('start_date', '>', $currentDate)
+                    ->orWhere('end_date', '<', $currentDate);
+            })
+            ->orderBy('end_date', 'desc');
     }
 
     public static function setAllMandatesExpiredInOrganization($organization) {