+++ csatar-plugins/csatar/updates/builder_table_update_csatar_csatar_scouts_4.php
@@ -3,21 +3,21 @@
 use Schema;
 use October\Rain\Database\Updates\Migration;
 
-class BuilderTableUpdateCsatarCsatarScouts3 extends Migration
+class BuilderTableUpdateCsatarCsatarScouts4 extends Migration
 {
     public function up()
     {
         Schema::table('csatar_csatar_scouts', function($table)
         {
-            $table->boolean('google_two_fa_is_activated')->after('google_two_fa_secret_key')->nullable();
+            $table->boolean('is_approved')->nullable()->default(false);
         });
     }
-    
+
     public function down()
     {
         Schema::table('csatar_csatar_scouts', function($table)
         {
-            $table->dropColumn('google_two_fa_is_activated');
+            $table->dropColumn('is_approved');
         });
     }
 }