+++ csatar-plugins/forms/traits/AjaxControllerSimple.php
@@ -108,6 +108,7 @@ trait AjaxControllerSimple {
             'form' => $html,
             'formUniqueId' => $this->formUniqueId,
             'additionalData' => $this->additionalData,
+            'specialValidationExceptions' => serialize($this->specialValidationExceptions),
             'recordKeyParam' => $this->recordKeyParam ?? Input::get('recordKeyParam'),
             'recordKeyValue' => $record->{$this->recordKeyParam ?? Input::get('recordKeyParam')} ?? 'new',
             'from_id' => $form->id,
@@ -621,6 +622,10 @@ trait AjaxControllerSimple {
             $attributeNames,
         );
 
+        if ($specialValidationExceptions = Input::get('specialValidationExceptions')) {
+            $specialValidationExceptions = unserialize($specialValidationExceptions);
+        }
+
         //validate for conditional rules
         if (isset($record->conditionalRules)) {
             foreach ($record->conditionalRules as $conditionalRule) {
@@ -630,7 +635,11 @@ trait AjaxControllerSimple {
             }
         }
 
-        if ($validation->fails()) {
+        if ($validation->fails() || !empty($specialValidationExceptions)) {
+            foreach ((array) $specialValidationExceptions as $key => $value) {
+                $validation->messages()->add('special_validation_exception_' . $key, $value);
+            }
+
             throw new \ValidationException($validation);
         }
 
@@ -1015,8 +1024,9 @@ trait AjaxControllerSimple {
                 if (array_key_exists('isPivot', $data)) {
                     $value = $relatedRecord->pivot->{$key} ?? '';
                 } else {
+                    $attribute = array_key_exists('valueFromFormBuilder', $data) ? $data['valueFromFormBuilder'] : 'name';
                     $value = (is_object($relatedRecord->{$key}) ?
-                        $relatedRecord->{$key}->name :
+                        $relatedRecord->{$key}->{$attribute} :
                         $relatedRecord->{$key});
                 }
                 $tableRows .= $value;