+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- CS-358 - Added design for the 2FA information display, added key icon to 2FA form-control fields
+- CS-441 - Added "Inactive mandates" accordion to Association, District, Team, Troop, Patrol pages
+- CS-443 - Changed accordion to tabs on organizations page
+- CS-424 - Hide inactive and expired mandates on Association, District, Team, Troop, Patrol pages
+- CS-424 - When listing mandates on frontend at organization data, the mandate owner name is clickable
 - CS-336 - Extended Personal Identification Number validation
 - CS-421 - Member card data import
 - CS-417 - Order team on district page by team number
@@ -9,6 +14,22 @@
 - CS-432 - Added "is_approved" field for Scout
 - CS-430 - Added option and menu to list all team reports in association regardless of team id
 - CS-418 - Create member card model
+- CS-356 - Added tags list to the patrol and troop forms
+
+## 1.7.2
+### 2023-01-18
+
+- CS-313 - Added Cancel button to the edit content form on the ogragization unit forms
+
+## 1.7.1
+### 2023-01-18
+
+- CS-448 - Fixed bug in team reports membership fee
+
+## 1.7.0
+### 2023-01-16
+
+- CS-440 - Fixed missing QR code for 2FA
 - CS-431 - separated inactive and active scouts on team page
 - CS-415 - inactive districts and teams are hidden in structure accordions and menu
 - CS-415 - inactive teams are hidden in district page